requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==2.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
PyYAML==3.11
SQLAlchemy==1.0.12
Werkzeug==0.11.4
